% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{error_packages}
\alias{error_packages}
\title{List of packages that failed to load in last scan}
\format{A character vector of package names}
\usage{
error_packages
}
\description{
While using \code{ls("package:pkgname")} to scan object names inside package
need to load and attach package first. Some packages cannot be loaded thus
cannot be scanned.
}
\details{
Some packages are simply not installed properly thus cannot be loaded.

In a big scan session some packages cannot be loadeded because of error
'maximal number of DLLs reached...'. The DLL limit is 100 according to
http://stackoverflow.com/questions/24832030/exceeded-maximum-number-of-dlls-in-r

The DLL limit was reached because too many packages were loaded in scan but
didn't get unloaded, although extra efforts was made to unload properly, some
still cannot be unloaded for dependency reason.

Running update_name_table(tryError = TRUE) in new R session can scan packages
that previously have problems. Each new scan in new R session should cover
some packages that failed in last time because of DLL limit error. After
several attempts all the left packages are packages have installation
problems. There is no need to scan again.

For example, there are 400 packages listed in package author's machine, 130
have errors in one scan from scratch, after 3 runs of scan error packages, 47
packages still left, all have installation problems.
}
\keyword{datasets}

