% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_names.R
\name{get_data_folder}
\alias{get_data_folder}
\title{Get path of installed_package_folder\\data\\}
\usage{
get_data_folder()
}
\value{
installed_package_folder\\data\\
}
\description{
Always load and save data to this folder to make sure only one version of
data exist
}

