% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{name_table}
\alias{name_table}
\title{data.table holding all scanned name result}
\format{A data.table with 2 columns:
\describe{
 \item{package}{Package name string}
 \item{obj_name}{Object name in package}
}}
\usage{
name_table
}
\description{
\code{ls("package:pkgname")} was used to scan all packages available.
Functions, datasets, operators, symbols, alternative formats like
\code{body()<-} are included from \code{ls()}.
}
\keyword{datasets}

