% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkg_list}
\alias{pkg_list}
\title{List of all installed packages from last scan}
\format{A character vector of package names}
\usage{
pkg_list
}
\description{
Saved when \code{pkg_name_changed()} was called.
\code{.packages(all.available = TRUE)} was used to check folder under library
location path \code{lib.loc}
}
\keyword{datasets}

