% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pkg_table}
\alias{pkg_table}
\title{data.table of all installed packages from last scan}
\format{A data.table with 3 columns:
\describe{
 \item{Package}{Package name string}
 \item{LibPath}{Package installation folder}
 \item{Version}{Version number string}
}}
\usage{
pkg_table
}
\description{
\code{installed.packages()} was used to check DESCRIPTION file for each package
folder, more accurate than checking name only.
}
\keyword{datasets}

