% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_builder.R
\name{println}
\alias{println}
\title{Helper method to print console message with default new line}
\usage{
println(...)
}
\arguments{
\item{...}{send to paste0}
}
\description{
Helper method to print console message with default new line
}

