% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_browser.R
\name{searchname_regex}
\alias{searchname_regex}
\title{Regex seach name in name table}
\usage{
searchname_regex()
}
\description{
Helper function to call \code{searchname} with regex enabled.
}
\details{
Need this because RStudio Addin currently don't support function with
parameters
}

