% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_browser.R
\name{searchnames}
\alias{searchnames}
\title{Seach name in name table}
\usage{
searchnames(search_regex = FALSE)
}
\arguments{
\item{search_regex}{}
}
\description{
With text selected or name around cursor as initial search term, find the
package and object name you need.
}
\details{
\code{Load Package} button run \code{library(package)} in console, insert
\code{library(package)} before current line, replace text selected or name
around cursor with selected object name. \code{Insert Package Prefix} button
replace text selected or name around cursor with selected object name
prefixed by object name.
}

