% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_index.R
\name{build_index}
\alias{build_index}
\title{Build name table}
\usage{
build_index()
}
\description{
All object names in all installed packages are scanned with
\code{ls(\"package:pkgname\")}.
}
\details{
Functions, datasets, operators, symbols, alternative formats like
\code{body()<-} are included from \code{ls()}. Package must be loaded and
attached first before using \code{ls()}. Thus all installed packages are
loaded and attached in the scanning process. Although extra efforts were made
to unload packages properly after use, there still will be some left over
when the scan finished. It's recommended to build index in a new R session
instead of working session with important data, and restart R session after
building.
}

