% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_builder.R
\name{pkg_name_changed}
\alias{pkg_name_changed}
\title{Scan package changes by name only}
\usage{
pkg_name_changed(startNew = FALSE)
}
\arguments{
\item{startNew}{Default FALSE, compare user's environment with name table
shipped with this package, only update difference. If True, build from
scratch.}
}
\value{
list(pkg_to_add, pkg_to_remove)
}
\description{
Compare currently packages name with previous list \code{pkg_list}.
}
\details{
Use \code{.packages(all.available = TRUE)} to check folder under library
location path \code{lib.loc}. Has more false positives but fast.
}

