% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_names.R
\name{pkg_name_changed}
\alias{pkg_name_changed}
\title{Scan package changes by name only}
\usage{
pkg_name_changed()
}
\value{
list(pkg_to_add, pkg_to_remove)
}
\description{
Compare currently packages name with previous list \code{pkg_list}.
}
\details{
Use \code{.packages(all.available = TRUE)} to check folder under library
location path \code{lib.loc}. Has more false positives but fast.
}

