% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_names.R
\name{pkg_name_version_changed}
\alias{pkg_name_version_changed}
\title{Scan package changes by name and version}
\usage{
pkg_name_version_changed()
}
\value{
list(pkg_to_add, pkg_to_remove)
}
\description{
Compare current packages name and version with previous table
\code{pkg_table}.
}
\details{
Use \code{installed.packages} to check DESCRIPTION file for each package folder, more accurate but slower than checking name only.
}

