% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_builder.R
\name{pkg_name_version_changed}
\alias{pkg_name_version_changed}
\title{Scan package changes by name and version}
\usage{
pkg_name_version_changed(startNew = FALSE)
}
\arguments{
\item{startNew}{Default FALSE, compare user's environment with name table
shipped with this package, only update difference. If True, build from
scratch.}
}
\value{
list(pkg_to_add, pkg_to_remove)
}
\description{
Compare current packages name and version with previous table
\code{pkg_table}.
}
\details{
Use \code{installed.packages()} to check DESCRIPTION file for each package
folder, more accurate than checking name only. R help cautioned it be slow if
thouands of packages available, but checking 300 ~ 400 packages didn't have
significant performance difference. It's recommended to always use this unless
it's too slow for you.
}

