% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_names.R
\name{update_name_table}
\alias{update_name_table}
\title{Update name table}
\usage{
update_name_table(withVersion = FALSE)
}
\arguments{
\item{withVersion}{If TRUE, update name table by changes both in name and
version version}
}
\description{
Update name table by package name changes, or by changes both in name and
version.
}
\details{
Use \code{.packages(all.available = TRUE)} for name changes,
\code{installed.packages} for name and version changes but slower.

In one case, \code{.packages(all.available = TRUE)} found 408 packages
folder, \code{installed.packages} found 379 packages with valid DESCRIPTION
file, the final loading, attaching, listing names function found 267 packages
with at least one name.
}

