% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_builder.R
\name{update_name_table}
\alias{update_name_table}
\title{Update name table}
\usage{
update_name_table(withVersion = TRUE, startNew = FALSE, tryError = FALSE)
}
\arguments{
\item{withVersion}{Default TRUE, update name table by changes both in name
and version. If FALSE, update by package name changes, a little bit faster
but with more false positives.}

\item{startNew}{Default FALSE, compare user's environment with name table
shipped with this package, only update difference. If True, build from
scratch.}

\item{tryError}{Default FALSE. If True, withVersion and startNew must be
FALSE, Scan the packages cannot be loaded in last update again.}
}
\description{
Update name table by package name changes, or by changes both in name and
version.
}
\details{
In one case, \code{.packages(all.available = TRUE)} found 408 packages
folder, \code{installed.packages} found 379 packages with valid DESCRIPTION
file, the final loading, attaching, listing names function found 267 packages
with at least one name.

This package shipped with a name table of these 267 packages. When user
updated name table in first time, the package changes were based on this name
table. User can also manually build name table from scratch by sending all
package names to
}

